重新游戏修复，二阶段快感知修复

+++ Assets/Scripts/UI/Play/PlayCanvas.cs
@@ -260,7 +260,7 @@ namespace act.ui
             processChangeSequence.AppendInterval(processDuration * 0.5f);
 
             processChangeSequence.Append(text_Process_Effect_Num.DOColor(new Color(color.r, color.g, color.b, 0.0f), processDuration));
-            processChangeSequence.Join(material_Process_Num.DOFloat(progress / 100, "_Progress", processDuration)
+            processChangeSequence.Join(material_Process_Num.DOFloat((float)progress / 100, "_Progress", processDuration)
                 .OnUpdate(() => {
                     //Debug.Log(material_Process_Num.GetFloat("_Progress"));
                     //Debug.Log(Convert.ToInt32((material_Process_Num.GetFloat("_Progress") * 100.0f)));
